pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "remi-dashboard-web"
version = "0.4.0"
description = "WZRD.dev Web Dashboard - Monitor and manage Remi and project agents"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [
    {name = "Joey Rodriguez"},
]
keywords = ["ai", "agent", "dashboard", "webui", "wzrd"]
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "pyyaml>=6.0",
    "watchfiles>=0.24.0",
]

[project.optional-dependencies]
tui = ["hermes-hud>=0.5.0"]
chat = ["hermes-agent>=0.1.0"]

[project.scripts]
remi-dashboard = "backend.main:cli"

[tool.setuptools.packages.find]
include = ["backend*"]

[tool.setuptools.package-data]
backend = ["static/**/*"]
